First of all added aws s3 integration Also added Bus status, bus parts repairman

+++ src/main/java/com/snob/busmanagmenttool/conroller/AdminController.java
@@ -1,14 +1,21 @@
 package com.snob.busmanagmenttool.conroller;
 
 import com.snob.busmanagmenttool.exception.EntityNotFoundException;
+import com.snob.busmanagmenttool.model.dto.BusDTO;
 import com.snob.busmanagmenttool.model.dto.CityDTO;
 import com.snob.busmanagmenttool.model.dto.UserDTO;
+import com.snob.busmanagmenttool.model.entity.machinery.Bus;
 import com.snob.busmanagmenttool.service.BusService;
 import com.snob.busmanagmenttool.service.CityService;
 import com.snob.busmanagmenttool.service.TicketService;
+import com.snob.busmanagmenttool.service.aws.S3Service;
 import com.snob.busmanagmenttool.service.user.BusDriverService;
 import com.snob.busmanagmenttool.service.user.UserService;
+
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import com.snob.busmanagmenttool.service.user.actions.FeedbackService;
@@ -18,6 +25,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @RestController
@@ -31,6 +39,7 @@ public class AdminController {
     private final BusDriverService busDriverService;
     private final FeedbackService feedbackService;
     private final TicketService ticketService;
+    private final S3Service s3Service;
 
 
     @GetMapping("/users")
@@ -43,11 +52,26 @@ public class AdminController {
     public Optional<UserDTO> getUser(@PathVariable Long id) throws EntityNotFoundException{
         return userService.getUserById(id);
     }
-    @DeleteMapping("/buses/{id}")
-    @PreAuthorize("hasAuthority('admin:delete')")
-    public void deleteBus(@PathVariable Long id) throws EntityNotFoundException {
+
+  @DeleteMapping("/buses/{id}")
+  @PreAuthorize("hasAuthority('admin:delete')")
+  public ResponseEntity<String> deleteBus(@PathVariable Long id) throws EntityNotFoundException {
+    try {
+      Optional<BusDTO> optionalBus = busService.getBusById(id);
+      if (optionalBus.isPresent()) {
+        BusDTO bus = optionalBus.get();
+        s3Service.deleteFile(bus.getPhotoUrl());
         busService.deleteBusById(id);
+        return ResponseEntity.ok("Bus deleted successfully.");
+      } else {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Bus not found.");
+      }
+    } catch (Exception e) {
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+          .body("Error deleting bus and image: " + e.getMessage());
     }
+    }
+
     @PostMapping("/city")
     @PreAuthorize("hasAuthority('admin:create')")
     public ResponseEntity<String> addCity(@RequestBody CityDTO cityDTO) {
@@ -76,4 +100,9 @@ public class AdminController {
         ticketService.deleteTicket(id);
         return ResponseEntity.noContent().build();
     }
+//    @PostMapping("/upload-image")
+//    public ResponseEntity<String> uploadBusImage(@RequestParam("file") MultipartFile file) throws IOException {
+//        String photoUrl = busImageService.uploadBusImage(file);
+//        return new ResponseEntity<>(photoUrl, HttpStatus.OK);
+//    }
 }
